Add AZP packaging of OSX
The conda build provide compiler requires an older version of the OSX
SDK. This is download and conda build is informed to use this sysroot
by conda configuration files. The CMake variables and command line
options are ignore by the conda toolchain, and the CONDA_BUILD_SYSROOT
has the highest priority.